Added configurations

+++ Project.Core/Models/IncomeDetail.cs
@@ -12,7 +12,7 @@ namespace Project.Core.Models
         public string IncomeName { get; set; }
         public decimal Price { get; set; }
         public string Exchange { get; set; }
-        public double MoneyOfTheDay { get; set; }
+        public decimal MoneyOfTheDay { get; set; }
         public string PaymentMethod { get; set; }
         public string Description { get; set; }
     }
